updates perch to 2.7.10

+++ cms/core/apps/content/PerchContent_PageTemplate.class.php
@@ -13,12 +13,24 @@ class PerchContent_PageTemplate extends PerchBase
     public function delete()
     {
     	$file = PerchUtil::file_path(PERCH_TEMPLATE_PATH.'/pages/'.$this->templatePath());
-    	if (file_exists($file)) {
+    	if (!PERCH_RUNWAY && file_exists($file)) {
     		unlink($file);
     	}
         parent::delete();
     }
 
-}
+    public function display_name() 
+    {
+        $out = '';
 
-?>
+        if (strpos($this->templatePath(), '/')!==false) {
+            $segments = explode('/', $this->templatePath());
+            array_pop($segments);
+            $out .= PerchUtil::filename(implode('/', $segments)).' → ';
+        }
+
+        $out .= $this->templateTitle();
+
+        return $out;
+    }
+}
